Migrated reference JS code to help system

+++ src/help/lib/tools/help.rb
@@ -8,10 +8,11 @@ module Blitz3D
 
         outdir = File.expand_path('../../../../_release/help', __dir__)
 
-        FileUtils.rm_r "#{outdir}/assets"
+        FileUtils.rm_r "#{outdir}/assets" if Dir.exist?("#{outdir}/assets")
         FileUtils.mkdir_p "#{outdir}/assets"
 
         File.write("#{outdir}/assets/style.css", Sass.compile(File.expand_path('../../assets/style.scss', __dir__), style: :compressed).css)
+        File.write("#{outdir}/assets/reference.js", Terser.compile(File.read(File.expand_path('../../assets/reference.js', __dir__)), compress: true))
 
         Blitz3D::Help::Index.new.generate("#{outdir}/index.html")
         Blitz3D::Help::Credits.new.generate("#{outdir}/credits.html")
